--- src/main/java/com/xmomen/module/base/company/mapper/CompanyMapper.java
@@ -1,6 +1,0 @@
-package com.xmomen.module.base.company.mapper;
-
-public interface CompanyMapper {
-    public static final String CompanyMapperNameSpace = "com.xmomen.module.base.company.mapper.CompanyMapper.";
-
-}
