+++ src/main/java/com/xmomen/module/base/mapper/CompanyMapper.java
@@ -1,0 +1,6 @@
+package com.xmomen.module.base.mapper;
+
+public interface CompanyMapper {
+    public static final String CompanyMapperNameSpace = "com.xmomen.module.base.mapper.CompanyMapper.";
+
+}
