+++ src/main/java/com/xmomen/module/base/mapper/CompanyMapper.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="com.xmomen.module.base.company.mapper.CompanyMapper" >
+<mapper namespace="com.xmomen.module.base.mapper.CompanyMapper" >
 
     <!--    查询消息    -->
-    <select id="getCompanyList" resultType="com.xmomen.module.base.company.model.CompanyModel" parameterType="java.util.HashMap">
+    <select id="getCompanyList" resultType="com.xmomen.module.base.model.CompanyModel" parameterType="java.util.HashMap">
         SELECT
         *
         FROM
