+++ src/main/java/com/xmomen/module/base/mapper/MemberMapper.java
@@ -1,0 +1,5 @@
+package com.xmomen.module.base.mapper;
+
+public interface MemberMapper {
+    public static final String MemberMapperNameSpace = "com.xmomen.module.base.mapper.MemberMapper.";
+}
