+++ src/main/java/com/xmomen/module/base/mapper/MemberMapper.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="com.xmomen.module.base.member.mapper.MemberMapper" >
+<mapper namespace="com.xmomen.module.base.mapper.MemberMapper" >
 
     <!--    查询消息    -->
-    <select id="getMemberList" resultType="com.xmomen.module.base.member.model.MemberModel" parameterType="java.util.HashMap">
+    <select id="getMemberList" resultType="com.xmomen.module.base.model.MemberModel" parameterType="java.util.HashMap">
         SELECT
         *
         FROM
