--- src/main/java/com/xmomen/module/base/member/mapper/MemberMapper.java
@@ -1,5 +1,0 @@
-package com.xmomen.module.base.member.mapper;
-
-public interface MemberMapper {
-    public static final String MemberMapperNameSpace = "com.xmomen.module.base.member.mapper.MemberMapper.";
-}
