+++ src/main/java/com/xmomen/module/base/service/CompanyService.java
@@ -1,7 +1,7 @@
-package com.xmomen.module.base.company.service;
+package com.xmomen.module.base.service;
 
-import com.xmomen.module.base.company.model.CreateCompany;
-import com.xmomen.module.base.company.model.UpdateCompany;
+import com.xmomen.module.base.model.CreateCompany;
+import com.xmomen.module.base.model.UpdateCompany;
 
 public interface CompanyService {
 	public void createCompany(CreateCompany createCompany);
