+++ src/main/java/com/xmomen/module/base/service/MemberSercvice.java
@@ -1,7 +1,7 @@
-package com.xmomen.module.base.member.service;
+package com.xmomen.module.base.service;
 
-import com.xmomen.module.base.member.model.CreateMember;
-import com.xmomen.module.base.member.model.UpdateMember;
+import com.xmomen.module.base.model.CreateMember;
+import com.xmomen.module.base.model.UpdateMember;
 
 public interface MemberSercvice {
 	public void createMember(CreateMember createMember);
